refactor: DataManager 타입 변경에 따른 test 파일명 변경
