+++ BoxOffice/Model/DataManager.swift
@@ -7,8 +7,8 @@
 
 import UIKit
 
-enum DataManager {
-    static func parse<T: Decodable>(from data: Data, returnType: T.Type) throws -> T? {
+struct DataManager {
+    func parse<T: Decodable>(from data: Data, returnType: T.Type) throws -> T? {
         let decoder = JSONDecoder()
         
         do {
@@ -16,21 +16,6 @@ enum DataManager {
             return result
         } catch {
             throw DecodeError.decodeFail
-        }
-    }
-}
-
-enum DecodeError: Error {
-    case decodeFail
-    
-    static let title = NSLocalizedString("오류 발생", comment: "error")
-}
-
-extension DecodeError: LocalizedError {
-    var errorDescription: String? {
-        switch self {
-        case .decodeFail:
-            return NSLocalizedString("데이터를 불러오는데 실패했습니다.", comment: "data loading failure")
         }
     }
 }
